Added test for MFAnalyser().getHTMLMolecularFormulaWithCharge().
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@4115 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/tools/MFAnalyser.java
@@ -615,7 +615,7 @@ public class MFAnalyser {
 	 */
 	public String getHTMLMolecularFormulaWithCharge() {
 		String formula = new MFAnalyser(atomContainer).getHTMLMolecularFormula();
-		double charge = (new AtomContainerManipulator()).getTotalFormalCharge(atomContainer);
+		int charge = (new AtomContainerManipulator()).getTotalFormalCharge(atomContainer);
 		if (charge == 0)
 		{
 			return formula;
